django-webapp/requirements.txt
# Минимальные зависимости для Django веб-приложения
# Без ML библиотек (PyTorch, torchvision)

Django>=4.2
Pillow>=10.0.0
onnxruntime>=1.16.0
onnx>=1.15.0
numpy>=1.24.0
boto3>=1.28.0
django-storages>=1.14.0
requests>=2.31.0
